Multiple Additions
Made it so that you can click on the title in the daily board and it will show you everything.

Made it so that the start and end locations can be clicked on which links to google maps.

Made it so that when you click + on an assignment in daily board that it automatically has the job displayed that you clicked it on.