Overhauled Loading and Unloading Locations, Added Backhaul Option With Corresponding Loading and Unloading Options, Added Field For Job Contractor Name and Contact
This should be the finished job creation form, there are some things that need to be fixed like allignment issues with the stepper at start of #4 till line 550 in the code (need to realign the thing). Will need to make adjustments to the spacing of fields and buttons as it looks a little packed with information right now.